ngubah modifier jadi atribut game object
