+++ modules/Menu.py
@@ -9,7 +9,7 @@ GAME_BG = (1, 0, 6)
 class Menu(ABC):
 
     click_sound = pygame.mixer.Sound("assets/sounds/click.wav")
-    click_sound.set_volume(0.5)
+    click_sound.set_volume(0.2)
 
     def __init__(self, name):
         self.name = name
@@ -131,17 +131,17 @@ def change_menu(menu:dict, crnt_page:str, screen):
             Menu.click_sound.play()
             return "play_menu"
 
-        elif menu["main_menu"].about_btn.check(pygame.mouse.get_pos()):
+        if menu["main_menu"].about_btn.check(pygame.mouse.get_pos()):
             menu["about_menu"].render(screen)
             Menu.click_sound.play()
             return "about_menu"
 
-        elif menu["main_menu"].info_btn.check(pygame.mouse.get_pos()):
+        if menu["main_menu"].info_btn.check(pygame.mouse.get_pos()):
             menu["info_menu"].render(screen)
             Menu.click_sound.play()
             return "info_menu"
 
-        elif menu["main_menu"].exit_btn.check(pygame.mouse.get_pos()):
+        if menu["main_menu"].exit_btn.check(pygame.mouse.get_pos()):
             Menu.click_sound.play()
             return "<exit>"
